Split SMT interface of counterexamples from checker
