+++ liquid-fixpoint.cabal
@@ -58,6 +58,7 @@ library
                     Language.Fixpoint.Counterexample
                     Language.Fixpoint.Counterexample.Build
                     Language.Fixpoint.Counterexample.Check
+                    Language.Fixpoint.Counterexample.SMT
                     Language.Fixpoint.Counterexample.Types
                     Language.Fixpoint.Defunctionalize
                     Language.Fixpoint.Graph
